Finished breaking down activity editor subscriptions
