+++ packages/twenty-front/src/modules/activities/components/ActivityComments.tsx
@@ -4,7 +4,6 @@ import { useRecoilValue } from 'recoil';
 import { v4 } from 'uuid';
 
 import { Comment } from '@/activities/comment/Comment';
-import { Activity } from '@/activities/types/Activity';
 import { Comment as CommentType } from '@/activities/types/Comment';
 import { currentWorkspaceMemberState } from '@/auth/states/currentWorkspaceMemberState';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
@@ -50,12 +49,12 @@ const StyledThreadCommentTitle = styled.div`
 `;
 
 type ActivityCommentsProps = {
-  activity: Pick<Activity, 'id'>;
+  activityId: string;
   scrollableContainerRef: React.RefObject<HTMLDivElement>;
 };
 
 export const ActivityComments = ({
-  activity,
+  activityId,
   scrollableContainerRef,
 }: ActivityCommentsProps) => {
   const { createOneRecord: createOneComment } = useCreateOneRecord({
@@ -66,10 +65,10 @@ export const ActivityComments = ({
 
   const { records: comments } = useFindManyRecords({
     objectNameSingular: CoreObjectNameSingular.Comment,
-    skip: !isNonEmptyString(activity?.id),
+    skip: !isNonEmptyString(activityId),
     filter: {
       activityId: {
-        eq: activity?.id ?? '',
+        eq: activityId,
       },
     },
   });
@@ -87,7 +86,7 @@ export const ActivityComments = ({
       id: v4(),
       authorId: currentWorkspaceMember?.id ?? '',
       author: currentWorkspaceMember,
-      activityId: activity?.id ?? '',
+      activityId: activityId,
       body: commentText,
       createdAt: new Date().toISOString(),
     });
