+++ packages/twenty-front/src/modules/activities/components/ActivityTypeDropdown.tsx
@@ -1,6 +1,7 @@
 import { useTheme } from '@emotion/react';
+import { useRecoilState } from 'recoil';
 
-import { Activity } from '@/activities/types/Activity';
+import { recordStoreFamilyState } from '@/object-record/record-store/states/recordStoreFamilyState';
 import {
   Chip,
   ChipAccent,
@@ -10,18 +11,21 @@ import {
 import { IconCheckbox, IconNotes } from '@/ui/display/icon';
 
 type ActivityTypeDropdownProps = {
-  activity: Pick<Activity, 'type'>;
+  activityId: string;
 };
 
 export const ActivityTypeDropdown = ({
-  activity,
+  activityId,
 }: ActivityTypeDropdownProps) => {
+  const [activityInStore] = useRecoilState(recordStoreFamilyState(activityId));
+
   const theme = useTheme();
+
   return (
     <Chip
-      label={activity.type}
+      label={activityInStore?.type}
       leftComponent={
-        activity.type === 'Note' ? (
+        activityInStore?.type === 'Note' ? (
           <IconNotes size={theme.icon.size.md} />
         ) : (
           <IconCheckbox size={theme.icon.size.md} />
